+++ App/Console/Console.php
@@ -1,0 +1,31 @@
+<?php
+namespace App\Console;
+
+use App\Console\Commands\Serve;
+use App\Console\Commands\BaseCommand;
+use Symfony\Component\Console\Application;
+
+class Console
+{
+    public function register()
+    {
+        return array(
+            Serve::class
+        );
+    }
+
+    /**
+     * @param Application $application
+     * @param BaseCommand $command
+     * @return bool
+     */
+    public function registerCommand(Application $application, BaseCommand $command)
+    {
+        if (!$command->isValid()) {
+            return false;
+        }
+
+        $application->add($command);
+        return $application;
+    }
+}